Fixing correctness

+++ examples/image_stitching_cuda/image_stitching.cu
@@ -37,10 +37,10 @@ __global__ void kernelWarpPerspective(int png_width, int png_height, int curr_wi
     int res_00 = (int)(prod_00/prod_20);
     int res_10 = (int)(prod_10/prod_20);
     if(res_00 >= 0 && res_00 < curr_width && res_10 >= 0 && res_10 < curr_height){
-        out_r_device[res_10*curr_width+res_00] = png_r[i*png_width + j]; 
-        out_g_device[res_10*curr_width+res_00] = png_g[i*png_width + j]; 
-        out_b_device[res_10*curr_width+res_00] = png_b[i*png_width + j];
-        out_a_device[res_10*curr_width+res_00] = png_a[i*png_width + j]; 
+        out_r_device[res_10*curr_width+res_00] = (int)png_r[i*png_width + j]; 
+        out_g_device[res_10*curr_width+res_00] = (int)png_g[i*png_width + j]; 
+        out_b_device[res_10*curr_width+res_00] = (int)png_b[i*png_width + j];
+        out_a_device[res_10*curr_width+res_00] = (int)png_a[i*png_width + j]; 
     }
     
 }
